+++ packages/apps/emails/src/functios-lib/get-mail-render.ts
@@ -1,9 +1,0 @@
-import transactionalMails from '@ecomplus/transactional-mails';
-
-const getMailRender = (templateName: string): ((...args: any[]) => Promise<string>) => {
-  const $render = global.$transactionalMails?.[templateName];
-  if (typeof $render === 'function') return $render;
-  return transactionalMails[templateName];
-};
-
-export default getMailRender;