+++ packages/apps/pagarme/src/pagarme.ts
@@ -3,10 +3,10 @@ import type { AppModuleBody } from '@cloudcommerce/types';
 import handleListPayments from './pagarme-list-payments';
 import handleCreateTransaction from './pagarme-create-transaction';
 
-export const listPayments = async (modBody: AppModuleBody) => {
+export const listPayments = async (modBody: AppModuleBody<'list_payments'>) => {
   return handleListPayments(modBody);
 };
 
-export const createTransaction = async (modBody: AppModuleBody) => {
+export const createTransaction = async (modBody: AppModuleBody<'create_transaction'>) => {
   return handleCreateTransaction(modBody);
 };