+++ packages/storefront/client.d.ts
@@ -25,7 +25,7 @@ interface Window {
   ECOM_CURRENCY: string;
   ECOM_CURRENCY_SYMBOL: string;
   ECOM_COUNTRY_CODE: string;
-  OAUTH_PROVIDERS?: Array<'google'>;
+  OAUTH_PROVIDERS?: Array<'google' | 'facebook'>;
   GIT_BRANCH: string;
   GIT_REPO?: string;
   AB_EXPERIMENT_ID?: string;